NewMainPage Draft
Alpha for the Main login page, demonstrating general format, if it's acceptable, then I'll move on to coordinate and beautify. Grey portions can be replaced with royalty-free imagery, static colors, or a gradient if desired. I recommend a royalty-free image. Login portion will be recolored to a more subdued and professional appearance.