Added side menu items + logout process
User will now populate a few template sidenav items (home, profile, admin, logout)

When the user logs out, the access token is revoked (added to the NoNoTokens table, and sends the response for the browser to clear the jwt token. If we don't clear the token, the user would receive an error stating the token is revoked.

TODO, i'd like for after some time 10 seconds? redirect the user back to the login page (from the loggedOut.html page.